fix: review style modifications
