+++ Cargo.toml
@@ -5,4 +5,4 @@ members = [
     "netcdf-src",
     "cftime",
 ]
-default-members = ["netcdf", "netcdf-sys"]
+default-members = ["netcdf", "netcdf-sys","cftime"]
