+++ cftime/src/datetimes/prolecpticgregorian.rs
@@ -1,47 +1,44 @@
+#![allow(unused)]
+
 use crate::calendars::Calendars;
 use crate::constants;
 use crate::durations::CFDuration;
+use crate::macros::{impl_date_display, impl_dt_display, impl_getter};
 use crate::time::Time;
 use crate::traits::{DateLike, DateTimeLike, IsLeap};
 use crate::tz::Tz;
-use crate::{impl_date_display, impl_dt_display, impl_getter};
-use chrono::Datelike;
 use num_integer::div_mod_floor;
 use std::{
     fmt,
     ops::{Add, Sub},
 };
-
 #[derive(Debug, Copy, Clone, Default)]
-pub struct DateAllLeap {
+pub struct DateProlepticGregorian {
     pub year: i32,
     pub month: u32,
     pub day: u32,
 }
-impl DateAllLeap {
+
+impl DateProlepticGregorian {
+    const DAYS_PER_MONTH: [u32; 12] = constants::DAYS_PER_MONTH;
+    const CUM_DAYS_PER_MONTH: [u32; 13] = constants::CUM_DAYS_PER_MONTH;
     const DAYS_PER_MONTH_LEAP: [u32; 12] = constants::DAYS_PER_MONTH_LEAP;
     const CUM_DAYS_PER_MONTH_LEAP: [u32; 13] = constants::CUM_DAYS_PER_MONTH_LEAP;
-    const CALENDAR: Calendars = Calendars::AllLeap;
-}
-#[allow(unused_variables)]
-impl IsLeap for DateAllLeap {
-    fn is_leap(year: i32) -> bool {
-        true
-    }
-}
+    const CALENDAR: Calendars = Calendars::ProlepticGregorian;
 
-impl DateAllLeap {
-    pub fn new(year: i32, month: u32, day: u32) -> DateAllLeap {
+    pub fn new(year: i32, month: u32, day: u32) -> DateProlepticGregorian {
         if (month > 12) | (month < 1) {
             panic!("Month should be between 1 and 12. Found {month}")
         }
-        let max_day = DateAllLeap::DAYS_PER_MONTH_LEAP[(month - 1) as usize];
-
+        let max_day = match DateProlepticGregorian::is_leap(year) {
+            true => DateProlepticGregorian::DAYS_PER_MONTH_LEAP[(month - 1) as usize],
+            false => DateProlepticGregorian::DAYS_PER_MONTH[(month - 1) as usize],
+        };
         if day > max_day {
             panic!(
                 "Day can not exceed {max_day} for {} of the year {year} and {}",
                 constants::MONTHS[(month - 1) as usize],
-                DateAllLeap::CALENDAR
+                DateProlepticGregorian::CALENDAR
             )
         }
         Self {
@@ -51,13 +48,23 @@ impl DateAllLeap {
         }
     }
 }
-impl DateLike for DateAllLeap {
+impl DateLike for DateProlepticGregorian {
     fn num_days_from_ce(&self) -> i32 {
-        let mut days: i32 = 0;
-        days += (self.year - constants::UNIX_DEFAULT_YEAR) * 366;
-        days += constants::CUM_DAYS_PER_MONTH_LEAP[(self.month - 1) as usize] as i32;
-        days += (self.day - 1) as i32;
-        days
+        let dy = self.year - constants::UNIX_DEFAULT_YEAR;
+        let mut num_days: u32 = (dy.abs() * 365) as u32;
+        for i in 0..dy.abs() {
+            if DateProlepticGregorian::is_leap(constants::UNIX_DEFAULT_YEAR + i) {
+                num_days += 1;
+            }
+        }
+        if DateProlepticGregorian::is_leap(constants::UNIX_DEFAULT_YEAR) {
+            num_days += constants::CUM_DAYS_PER_MONTH_LEAP[(self.month - 1) as usize]
+        } else {
+            num_days += constants::CUM_DAYS_PER_MONTH[(self.month - 1) as usize]
+        }
+        // First begin to 1
+        num_days += self.day - 1;
+        (num_days as i32) * dy.signum()
     }
     fn num_hours_from_ce(&self) -> i32 {
         self.num_days_from_ce() * 24
@@ -72,48 +79,93 @@ impl DateLike for DateAllLeap {
         ((self.num_seconds_from_ce() as f64) * 1e6) as i64
     }
 
-    fn from_timestamp(seconds: i32) -> DateAllLeap {
-        let (nb_days, _) = div_mod_floor(seconds, constants::SECS_PER_DAY as i32);
-        let (nb_year, mut remaining_days) = div_mod_floor(nb_days, 366);
+    fn from_timestamp(seconds: i32) -> DateProlepticGregorian {
+        let (nb_days, seconds) = div_mod_floor(seconds, constants::SECS_PER_DAY as i32);
 
-        remaining_days += constants::UNIX_DEFAULT_DAY as i32;
-        let mut month: u32 = constants::UNIX_DEFAULT_MONTH;
-        for v in DateAllLeap::DAYS_PER_MONTH_LEAP.iter() {
-            if remaining_days - (*v as i32) >= 0 && ((month + 1) <= 12) {
-                remaining_days -= *v as i32;
-                month += 1
-            } else {
-                break;
+        let nb_non_leap_years = nb_days / 365;
+
+        let mut year = constants::UNIX_DEFAULT_YEAR + nb_non_leap_years;
+
+        let mut year_start = constants::UNIX_DEFAULT_YEAR;
+        let mut year_end = constants::UNIX_DEFAULT_YEAR + nb_non_leap_years;
+
+        let mut nb_of_leap_year = 0;
+        if year_end < year_start {
+            (year_start, year_end) = (year_end, year_start);
+        }
+
+        for year in year_start..year_end {
+            if DateProlepticGregorian::is_leap(year) {
+                nb_of_leap_year += 1;
             }
         }
-        DateAllLeap::new(
-            constants::UNIX_DEFAULT_YEAR + nb_year,
-            month,
-            (remaining_days) as u32,
-        )
+        let mut remaining_days = nb_days % 365;
+        if remaining_days.is_negative() {
+            year -= 1;
+            // remaining_days += 365 + nb_of_leap_year
+            remaining_days = remaining_days + 365 + nb_of_leap_year + 1
+        } else {
+            remaining_days = remaining_days - nb_of_leap_year + 1
+        }
+        // if 365 then 0;
+        // Do not include 00:00:00 as previous day
+        if seconds == 0 && nb_days.is_negative() && DateProlepticGregorian::is_leap(year) {
+            remaining_days += 1;
+        }
+
+        let mut month: u32 = 1;
+        if DateProlepticGregorian::is_leap(year) {
+            for v in DateProlepticGregorian::DAYS_PER_MONTH_LEAP.iter() {
+                if remaining_days - (*v as i32) >= 0 && ((month + 1) <= 12) {
+                    remaining_days -= *v as i32;
+                    month += 1
+                } else {
+                    break;
+                }
+            }
+        } else {
+            for v in DateProlepticGregorian::DAYS_PER_MONTH.iter() {
+                if remaining_days - (*v as i32) > 0 {
+                    remaining_days -= *v as i32;
+                    month += 1;
+                } else {
+                    break;
+                }
+            }
+        }
+        DateProlepticGregorian::new(year, month, remaining_days as u32)
+    }
+}
+impl IsLeap for DateProlepticGregorian {
+    fn is_leap(year: i32) -> bool {
+        let mut f_year = year;
+        if year < 0 {
+            f_year = year + 1;
+        }
+        (f_year % 400 == 0) | ((f_year % 4 == 0) && (f_year % 100 != 0))
     }
 }
 
 #[derive(Debug, Copy, Clone, Default)]
-pub struct DateTimeAllLeap {
-    pub date: DateAllLeap,
+pub struct DateTimeProlepticGregorian {
+    pub date: DateProlepticGregorian,
     pub time: Time,
     pub tz: Tz,
 }
 
-impl DateTimeAllLeap {
-    pub fn new(date: DateAllLeap, time: Time, tz: Tz) -> DateTimeAllLeap {
-        DateTimeAllLeap {
+impl DateTimeProlepticGregorian {
+    pub fn new(date: DateProlepticGregorian, time: Time, tz: Tz) -> DateTimeProlepticGregorian {
+        DateTimeProlepticGregorian {
             date: date,
             time: time,
             tz: tz,
         }
     }
 }
-impl DateTimeLike for DateTimeAllLeap {
+impl DateTimeLike for DateTimeProlepticGregorian {
     fn from_hms(hour: u32, minute: u32, second: u32) -> Self {
         Self {
-            date: DateAllLeap::new(
+            date: DateProlepticGregorian::new(
                 constants::UNIX_DEFAULT_YEAR,
                 constants::UNIX_DEFAULT_MONTH,
                 constants::UNIX_DEFAULT_DAY,
@@ -124,14 +176,14 @@ impl DateTimeLike for DateTimeAllLeap {
     }
     fn from_ymd(year: i32, month: u32, day: u32) -> Self {
         Self {
-            date: DateAllLeap::new(year, month, day),
+            date: DateProlepticGregorian::new(year, month, day),
             time: Time::new(0, 0, 0, 0),
             tz: Tz { hour: 0, minute: 0 },
         }
     }
     fn from_timestamp(seconds: i32) -> Self {
         Self {
-            date: DateAllLeap::from_timestamp(seconds),
+            date: DateProlepticGregorian::from_timestamp(seconds),
             time: Time::from_timestamp(seconds),
             tz: Tz { hour: 0, minute: 0 },
         }
@@ -163,11 +215,11 @@ fn _get_real_nano_field(duration: CFDuration) -> i64 {
     ns
 }
 
-impl Add<CFDuration> for DateTimeAllLeap {
+impl Add<CFDuration> for DateTimeProlepticGregorian {
     type Output = Self;
     fn add(self, other: CFDuration) -> Self {
         let ns = _get_real_nano_field(other);
-        let mut dt = DateTimeAllLeap::from_timestamp(
+        let mut dt = DateTimeProlepticGregorian::from_timestamp(
             self.num_seconds_from_ce() + other.num_seconds() as i32,
         );
         dt.time.nanosecond = ns as u64;
@@ -175,7 +227,7 @@ impl Add<CFDuration> for DateTimeAllLeap {
     }
 }
 
-impl Sub<CFDuration> for DateTimeAllLeap {
+impl Sub<CFDuration> for DateTimeProlepticGregorian {
     type Output = Self;
     fn sub(self, other: CFDuration) -> Self {
         let ns = _get_real_nano_field(other);
@@ -183,7 +235,7 @@ impl Sub<CFDuration> for DateTimeAllLeap {
         if ns > 0 {
             timestamp -= 1
         }
-        let mut dt = DateTimeAllLeap::from_timestamp(timestamp);
+        let mut dt = DateTimeProlepticGregorian::from_timestamp(timestamp);
         if ns > 0 {
             dt.time.nanosecond = (constants::MAX_NS - other.num_nanoseconds().unwrap()) as u64;
         }
@@ -191,9 +243,9 @@ impl Sub<CFDuration> for DateTimeAllLeap {
     }
 }
 
-impl_getter!(DateAllLeap);
-impl_date_display!(DateAllLeap);
-impl_dt_display!(DateTimeAllLeap);
+impl_getter!(DateProlepticGregorian);
+impl_date_display!(DateProlepticGregorian);
+impl_dt_display!(DateTimeProlepticGregorian);
 
 #[cfg(test)]
 mod test {
@@ -201,7 +253,7 @@ mod test {
     use crate::durations::CFDuration;
     #[test]
     fn test_add_duration_to_datetime() {
-        let dt = DateTimeAllLeap::from_timestamp(0);
+        let dt = DateTimeProlepticGregorian::from_timestamp(0);
         let dur = CFDuration::days(1, Calendars::ProlepticGregorian);
         let new_dt = dt + dur;
         assert_eq!(new_dt.date.year, 1970);
@@ -210,7 +262,7 @@ mod test {
         assert_eq!(new_dt.time.hour, 00);
         assert_eq!(new_dt.time.minute, 00);
         assert_eq!(new_dt.time.second, 00);
-        let dt = DateTimeAllLeap::from_timestamp(0);
+        let dt = DateTimeProlepticGregorian::from_timestamp(0);
         let dur = CFDuration::milliseconds(1, Calendars::ProlepticGregorian);
         let new_dt = dt + dur;
         assert_eq!(new_dt.date.year, 1970);
@@ -223,7 +275,7 @@ mod test {
     }
     #[test]
     fn test_sub_duration_to_datetime() {
-        let dt = DateTimeAllLeap::from_timestamp(0);
+        let dt = DateTimeProlepticGregorian::from_timestamp(0);
         let dur = CFDuration::days(1, Calendars::ProlepticGregorian);
         let new_dt = dt - dur;
         println!("{new_dt}");
@@ -233,7 +285,7 @@ mod test {
         assert_eq!(new_dt.time.hour, 00);
         assert_eq!(new_dt.time.minute, 00);
         assert_eq!(new_dt.time.second, 00);
-        let dt = DateTimeAllLeap::from_timestamp(0);
+        let dt = DateTimeProlepticGregorian::from_timestamp(0);
         let dur = CFDuration::milliseconds(1, Calendars::ProlepticGregorian);
         let new_dt = dt - dur;
         assert_eq!(new_dt.date.year, 1969);
@@ -246,8 +298,8 @@ mod test {
     }
 
     #[test]
-    fn test_from_timestam() {
-        let dt = DateTimeAllLeap::from_timestamp(0);
+    fn test_from_timestamp() {
+        let dt = DateTimeProlepticGregorian::from_timestamp(0);
         println!("{dt}");
         assert_eq!(dt.date.year, 1970);
         assert_eq!(dt.date.month, 01);
@@ -256,7 +308,7 @@ mod test {
         assert_eq!(dt.time.minute, 00);
         assert_eq!(dt.time.second, 00);
         // Bug found for this value
-        let dt = DateTimeAllLeap::from_timestamp(-86400);
+        let dt = DateTimeProlepticGregorian::from_timestamp(-86400);
         println!("{dt}");
         assert_eq!(dt.date.year, 1969);
         assert_eq!(dt.date.month, 12);
@@ -264,14 +316,14 @@ mod test {
         assert_eq!(dt.time.hour, 00);
         assert_eq!(dt.time.minute, 00);
         assert_eq!(dt.time.second, 00);
-        let dt = DateTimeAllLeap::from_timestamp(-1);
+        let dt = DateTimeProlepticGregorian::from_timestamp(-1);
         assert_eq!(dt.date.year, 1969);
         assert_eq!(dt.date.month, 12);
         assert_eq!(dt.date.day, 31);
         assert_eq!(dt.time.hour, 23);
         assert_eq!(dt.time.minute, 59);
         assert_eq!(dt.time.second, 59);
-        let dt = DateTimeAllLeap::from_timestamp(1000000);
+        let dt = DateTimeProlepticGregorian::from_timestamp(1000000);
         println!("{dt}");
         assert_eq!(dt.date.year, 1970);
         assert_eq!(dt.date.month, 01);
@@ -279,19 +331,19 @@ mod test {
         assert_eq!(dt.time.hour, 13);
         assert_eq!(dt.time.minute, 46);
         assert_eq!(dt.time.second, 40);
-        let dt = DateTimeAllLeap::from_timestamp(1658876523);
+        let dt = DateTimeProlepticGregorian::from_timestamp(1658876523);
         println!("{dt}");
         assert_eq!(dt.date.year, 2022);
-        assert_eq!(dt.date.month, 06);
-        assert_eq!(dt.date.day, 16);
+        assert_eq!(dt.date.month, 07);
+        assert_eq!(dt.date.day, 26);
         assert_eq!(dt.time.hour, 23);
         assert_eq!(dt.time.minute, 02);
         assert_eq!(dt.time.second, 03);
-        let dt = DateTimeAllLeap::from_timestamp(-1658876523);
+        let dt = DateTimeProlepticGregorian::from_timestamp(-1658876523);
         println!("{dt}");
         assert_eq!(dt.date.year, 1917);
-        assert_eq!(dt.date.month, 07);
-        assert_eq!(dt.date.day, 17);
+        assert_eq!(dt.date.month, 06);
+        assert_eq!(dt.date.day, 08);
         assert_eq!(dt.time.hour, 00);
         assert_eq!(dt.time.minute, 57);
         assert_eq!(dt.time.second, 57);
