+++ cftime/src/lib.rs
@@ -1,15 +1,11 @@
 // //! Time handling according to CF conventions
 // #![allow(missing_docs)]
-mod calendars;
-mod constants;
-mod datetime_360day;
-mod datetime_all_leap;
-mod datetime_julian;
-mod datetime_no_leap;
-mod datetime_prolecpticgregorian;
-mod durations;
-mod macros;
-mod parser;
-mod time;
-mod traits;
-mod tz;
+pub mod calendars;
+pub mod constants;
+pub mod datetimes;
+pub mod durations;
+pub mod macros;
+pub mod parser;
+pub mod time;
+pub mod traits;
+pub mod tz;
