+++ cftime/src/macros.rs
@@ -1,4 +1,3 @@
-#[macro_export]
 macro_rules! impl_getter {
     ($date:ident) => {
         impl $date {
@@ -15,7 +14,6 @@ macro_rules! impl_getter {
     };
 }
 
-#[macro_export]
 macro_rules! impl_date_display {
     ($date:ident) => {
         impl fmt::Display for $date {
@@ -31,7 +29,6 @@ macro_rules! impl_date_display {
     };
 }
 
-#[macro_export]
 macro_rules! impl_dt_display {
     ($datetime:ident) => {
         impl fmt::Display for $datetime {
@@ -46,3 +43,6 @@ macro_rules! impl_dt_display {
         }
     };
 }
+pub(crate) use impl_date_display;
+pub(crate) use impl_dt_display;
+pub(crate) use impl_getter;
