+++ cftime/src/traits.rs
@@ -1,8 +1,17 @@
+#![allow(unused)]
+use std::{
+    fmt::Debug,
+    ops::{Add, Sub},
+};
+
 use crate::{
     calendars::Calendars,
-    parser::{cf_parser, ParsedCFTime},
-    time::Time,
-    tz::Tz,
+    datetimes::{
+        day360::{Date360Day, DateTime360Day},
+        factory::CFDatetimes,
+    },
+    durations::CFDuration,
+    parser::cf_parser,
 };
 pub trait IsLeap {
     fn is_leap(year: i32) -> bool;
@@ -12,10 +21,10 @@ pub trait CFTimeEncoder {
     fn encode(unit: &str, calendar: Calendars);
 }
 pub trait CFTimeDecoder {
-    fn decode(&self, unit: &str, calendar: Option<Calendars>);
+    fn decode(self, unit: &str, calendar: Option<Calendars>);
 }
 
-pub trait DateLike {
+pub trait DateLike: Debug {
     fn num_days_from_ce(&self) -> i32;
     fn num_hours_from_ce(&self) -> i32;
     fn num_minutes_from_ce(&self) -> i32;
@@ -25,7 +34,7 @@ pub trait DateLike {
     where
         Self: Sized;
 }
-pub trait DateTimeLike {
+pub trait DateTimeLike: Debug {
     fn from_hms(hour: u32, minute: u32, second: u32) -> Self
     where
         Self: Sized;
@@ -39,11 +48,4 @@ pub trait DateTimeLike {
     fn num_minutes_from_ce(&self) -> i32;
     fn num_seconds_from_ce(&self) -> i32;
     fn num_nanoseconds_from_ce(&self) -> i64;
-}
-impl CFTimeDecoder for f64 {
-    fn decode(&self, unit: &str, calendar: Option<Calendars>) {
-        let parsed_cf_time = cf_parser(unit, calendar).unwrap();
-        let duration: f64 = parsed_cf_time.duration.num_seconds() as f64 * *self;
-        let from = parsed_cf_time.from;
-    }
 }
